Support Automatic ID Generation (#54)
ElasticSearch supports [automatically generating an ID for documents](https://www.elastic.co/guide/en/elasticsearch/reference/5.5/docs-index_.html#_automatic_id_generation) where id is not specified. To automatically generate an id with ElasticSearch the document must be sent as a `POST` request to the index.

This adds a condition to `ElasticRecord::Index::Documents#index_document` which will `POST` a document to the index if no id is specified.