Alter flow to understand update vs search
The previous fix was using ActiveRecord to determine behavior. Instead,
refactoring the code to understand if it's an update vs a simple search
seems like a better way to ensure we're doing the right thing. With this
change, we'll still be setting the nil key if it's an update as it will
be part of the changes, but we don't need to use provided? (nor mess
with as_search_document)