Make ElasticRecord ES 7.X Compatible (#103)
* Remove mapping type name

Mapping types are deprecated in ES7.

* Update results count

This now returns a hash of the value and rel.

* Make index mapping structure conditional

The old structure is going away but this addition makes it conditional
based on the presence of a custom mapping_type. Should work with 6.X and
7.X.

I noticed that if you don't explicitly set the include_type_name
parameter ES will get upset depending on the structure you used so at
least for 6.X and 7.X it seems better to include it explicitly even if
your mapping matches the format for that version/default param value.