Remove type param/routing in indexing docs (#104)
Deprecated/gone in 7+, not needed in 6. Since Data-Axle is already not
using and/or reliant on this, start stripping the rest of it out.

There's still some alternative uses that mapping_type serves for
has_child queries and such, so I can't completely remove all of it as
I'd like, but this will at least get rid of some observed ES7
deprecation warnings.