Ensure relations return full total of results
ElasticSearch introduced a new feature where a search has the total
limited to 10k by default for all requests. While this might make sense
for some usages, elastic_record is designed to be an ORM. You wouldn't
expect a SQL COUNT to lie to you unless you gave it a magic word.