Fix `joins` and `where` on ElasticRecord::Relation
`where` would fail if it was combined with `filter`. This happened
because filter would do its filtering by using the find command. Find
fails if an id is passed in that doesn't exist in the relation. In other
words `where` only worked with filter if where didn't remove anything.

I also added tests for the `joins` method.